+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Fixed & Added
+
+- New tests on process start
+- Create helper to modify settings in tests
+
+
 ## [Unreleased]
 ### Added
 - Dutch translations for interface